Commit: Fix MenuBar User State Handling
The MenuBar component has been modified to improve user state handling. It now uses the following logic in useState:
useState(rawToken ? jose.decodeJwt(rawToken) : user.toString());
This change ensures that if the user has the app open in two tabs and logs off in one of them, when trying to navigate in the other tab they will be automatically redirected to the login page.