Feature: Message Component
The validation of the message field when entering or editing a new message has been changed so that the entered value must be at least one letter long.
The validation was removed by commenting out the name property of the Form.Item Contacts field, as it is possible for a predefined message to exist without associated contacts.